MODIFY: User Validator and SignUpView
- User Validator 부분 User app의 utils.py로 분리
- email, password validator가 검증된 값 리턴하게 수정, 가독성 고려
- validate_duplicate 함수 가독성 향상
- 그 외 코드포매팅
