+++ users/utils.py
@@ -1,0 +1,24 @@
+from django.core.exceptions import ValidationError
+
+def validate_email(email):
+    if not ("@" in email and "." in email):
+        raise ValidationError(f"{email} is not an valid email.")
+    
+    return email
+
+def validate_password(password):
+    MIN_PASSWORD_LENGTH = 8
+
+    if len(str(password)) < MIN_PASSWORD_LENGTH:
+        raise ValidationError("Your Password is too short. Use Longer Password.")
+    
+    return password
+
+def validate_gender(gender):
+    if not gender:
+        return 'N'
+
+    if gender not in ['M', 'F']:
+        raise ValidationError("Gender Field should be either 'M' or 'F' or empty.")
+
+    return gender
