+++ users/views.py
@@ -1,5 +1,4 @@
 import bcrypt
-import jwt
 import json
 from json                   import JSONDecodeError
 
@@ -8,29 +7,29 @@ from django.http            import JsonResponse
 from django.core.exceptions import ValidationError
 
 from users.models           import User
-from utils.validators       import validate_email, validate_password, validate_gender, validate_duplicate, DuplicatedEntryError
+from users.utils            import validate_email, validate_password, validate_gender
+from utils.validators       import validate_duplicate, DuplicatedEntryError
 
 class SignUpView(View):
-
     def post(self, request):
 
         try:
             data = json.loads(request.body)
 
-            validate_email(data["email"])
-            validate_password(data["password"])
-            data['gender'] = validate_gender(data.get('gender'))
+            email = validate_email(data["email"])
+            password = validate_password(data["password"])
+            gender = validate_gender(data.get('gender'))
             validate_duplicate(User, data)
-
-            hashed_password = bcrypt.hashpw(str(data.get('password')).encode('utf-8'), bcrypt.gensalt()).decode('utf-8')
-
+            
+            hashed_password = bcrypt.hashpw(password.encode('utf-8'), bcrypt.gensalt()).decode('utf-8')
+            
             user = User.objects.create(
-                email           = data.get('email'),
+                email           = email,
                 password        = hashed_password,
                 phone_number    = data['phone_number'],
                 name            = data['name'],
                 birthday        = data.get('birthday'),
-                gender          = data.get('gender'),
+                gender          = gender,
             )
 
             return JsonResponse({"status": "SUCCESS", "user": user.to_dict()}, status=200)
